add managment exception to to user and calendar

+++ src/main/java/com/example/calendar/service/CalendarService.java
@@ -4,11 +4,14 @@ import com.example.calendar.model.Calendar;
 import com.example.calendar.model.Event;
 import com.example.calendar.model.User;
 import com.example.calendar.repository.CalendarRepository;
+import com.example.calendar.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 
@@ -18,46 +21,55 @@ public class CalendarService {
     @Autowired
     CalendarRepository calendarRepository;
 
-    public void createCalendar(Calendar calendar){
-        calendarRepository.save(calendar);
-    }
+    @Autowired
+    UserRepository userRepository;
 
-    public List<Calendar> getAllCalendar(){
-        return calendarRepository.findAll();
-    }
-
-    @Transactional
-    public void updateCalendar(int id,
-                               Optional<String> name,
-                               Optional<String> description,
-                               Optional<User> user) {
-
-        Calendar calendar = calendarRepository.getById(id);
-
-        if (calendar != null) {
-            name.ifPresent(calendar::setName);
-            description.ifPresent(calendar::setDescription);
-            user.ifPresent(calendar::setUser);
-
-            calendarRepository.updateCalendar(id,
-                    calendar.getName(),
-                    calendar.getDescription(),
-                    calendar.getUser());
+    public Calendar createCalendar(int userID, Calendar calendar) throws Exception{
+        if (userRepository.findById(userID).isPresent()){
+            return calendarRepository.save(calendar);
+        } else {
+            throw new Exception(String.format("User with ID %s not exist", userID));
         }
-
-    }
-    /*
-    public void updateCalendar(int id,Calendar newCalendar){
-
-        calendarRepository.updateCalendar(id, newCalendar.getName(), newCalendar.getDescription(), newCalendar.getUser());
-
-        Calendar oldCalendar = calendarRepository.findById(id).orElse(null);
     }
 
-     */
+    public List<Calendar> getAllCalendarByUser(int userID) throws Exception{
+        Optional<User> user = userRepository.findById(userID);
 
-    public void deleteCalendar(int id){
-        calendarRepository.deleteById(id);
+        if (user.isPresent()) {
+            return new ArrayList<>(user.get().getCalendars());
+        } else {
+            throw new Exception(String.format("User with ID %s don't have calendars", userID));
+        }
+    }
+
+
+    public Calendar updateCalendar(int id, Calendar updateCalendar) throws Exception{
+        if (calendarRepository.findById(id).isPresent()){
+
+            Calendar calendar = calendarRepository.findById(id).get();
+
+            if (Objects.nonNull(updateCalendar.getName())){
+                calendar.setName(updateCalendar.getName());
+            }
+
+            if (Objects.nonNull(updateCalendar.getDescription())){
+                calendar.setDescription(updateCalendar.getDescription());
+            }
+
+            return calendarRepository.save(calendar);
+        } else {
+            throw new Exception(String.format("Calendar with ID %s not found", id));
+        }
+    }
+
+
+    public String deleteCalendar(int id) throws Exception{
+        if (calendarRepository.findById(id).isPresent()){
+            calendarRepository.deleteById(id);
+            return String.format("Calendar with ID %s deleted", id);
+        } else {
+            throw new Exception(String.format("Calendar with ID %s not exist", id));
+        }
     }
 
 }
